Move library to root of repository
A popular library installation technique is to download the library via GitHub's Clone or download > Download ZIP and then use the Arduino IDE's Sketch > Include Library > Add .ZIP Library on the downloaded file. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

The previous use of a src folder indicated to me that the 1.5 Arduino Library format was preferred and so I have also added the required library.properties metadata file. This does break backwards compatibility with the seriously outdated Arduino IDE 1.0.x versions.

Reference: https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification

This move is also required if you wanted to add your library to the Arduino Library Manager index, which provides an even easier installation option.